requirements.txt
streamlit
cohere
python-dotenv
pdfplumber
python-docx
pytesseract
Pillow
langchain
faiss-cpu

